Implement DEVICE_SLEEP for RP2, implement DEVICE_LPTICKER for RP235x (#574)
* RP2xxx: Initial sleep implementation for RP2040, remove broken RTC LP ticker code

* Initial LPTICKER implementation for RP2350. Not tested yet.

* Fix missing interrupt constants, LPTICKER test working

* Add deep sleep latency

* Remove debug print, add missing file

* Cite source

* Update changelog

+++ cmsis/device/rtos/source/mbed_boot.c
@@ -139,7 +139,7 @@ static void mbed_cpy_nvic(void)
 #if defined(NVIC_RAM_VECTOR_ADDRESS)
     uint32_t *old_vectors = (uint32_t *)SCB->VTOR;
     uint32_t *vectors = (uint32_t *)NVIC_RAM_VECTOR_ADDRESS;
-    for (int i = 0; i < NVIC_NUM_VECTORS; i++) {
+    for (int i = 0; i < (int)(NVIC_NUM_VECTORS); i++) {
         vectors[i] = old_vectors[i];
     }
     SCB->VTOR = (uint32_t)NVIC_RAM_VECTOR_ADDRESS;
